+++ doc/examples/classic.py
@@ -3,7 +3,7 @@ import numpy as np
 # don't import any random modules in a Pythran file. Here, no problem!
 from fluiddyn.util import mpi
 
-from fluidpythran import pythran_def
+from fluidpythran import boost
 
 # FLUIDPYTHRAN_NO_IMPORT
 
@@ -13,7 +13,7 @@ from fluidpythran import pythran_def
 # pythran def func(int[][], float[][])
 
 
-@pythran_def
+@boost
 def func(a, b):
     return (a * np.log(b)).max()