+++ doc/examples/mixed_classic_type_hint.py
@@ -3,7 +3,7 @@ import numpy as np
 # don't import skimage in a Pythran file. Here, no problem!
 from skimage.filters import sobel
 
-from fluidpythran import pythran_def
+from fluidpythran import boost
 
 # pythran import numpy as np
 
@@ -11,11 +11,11 @@ from fluidpythran import pythran_def
 # pythran def func(int[][], float[][])
 
 
-@pythran_def
+@boost
 def func(a: float, b: float):
     return (a * np.log(b)).max()
 
 
-@pythran_def
+@boost
 def func1(a: int, b: float):
     return a * np.cos(b)