+++ doc/examples/type_hints_notemplate.py
@@ -1,5 +1,5 @@
 import numpy as np
-from fluidpythran import Type, NDim, Array, pythran_def
+from fluidpythran import Type, NDim, Array, boost
 
 T = Type(int, np.complex128)
 N = NDim(1, 3)
@@ -8,7 +8,7 @@ A = Array[T, N]
 A1 = Array[np.float32, N + 1]
 
 
-@pythran_def
+@boost
 def compute(a: A, b: A, c: T, d: A1, e: str):
     print(e)
     tmp = a + b